# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tole Community Management API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Tole Community Management System",
      "url": "https://github.com/"
    },
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "This server — all versioned endpoints live under the /api prefix."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `token` returned by `POST /auth/login` into the Authorize dialog. It is sent as `Authorization: Bearer <token>`. Query-string tokens are rejected by the server."
      }
    },
...
    "parameters": {
...
      "RoleQuery": {
        "name": "role",
        "in": "query",
        "required": false,
        "description": "Filter by role. **Note:** staff and resident callers are silently forced to `role=staff` with `isActive=true` regardless of what they send — only admins can list all roles.",
        "schema": {
          "$ref": "#/components/schemas/UserRole"
        }
      },
...
    },
    "responses": {
      "UnauthorizedResponse": {
        "description": "Authentication failed or was not supplied. Returned when the `Authorization` header is missing, the token is invalid or expired, the referenced user no longer exists, the account is deactivated, or a token was passed as a `token` / `access_token` **query parameter** (which this API rejects on purpose to avoid leaking credentials in logs).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "missingToken": {
            "summary": "No Authorization header",
            "value": {
              "success": false,
              "message": "Not authorized"
            }
          },
          "queryToken": {
            "summary": "Token supplied in the query string",
            "value": {
              "success": false,
              "message": "Use the Authorization request header"
            }
          },
          "inactive": {
            "summary": "Deactivated account",
            "value": {
              "success": false,
              "message": "User account is inactive or not found"
            }
          },
          "invalid": {
            "summary": "Malformed or bad signature",
            "value": {
              "success": false,
              "message": "Token invalid"
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "The caller is authenticated but not permitted. Covers role mismatches (`authorize`), record-level ownership checks (a resident touching another resident's due or complaint), and denied export attempts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "examples": {
          "role": {
            "summary": "Wrong role for this endpoint",
            "value": {
              "success": false,
              "message": "Access denied"
            }
          },
          "export": {
            "summary": "Export not permitted for this staff member",
            "value": {
              "success": false,
              "message": "You are not allowed to export dues data."
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "The resource does not exist. Also returned when a file is requested that is missing on disk, and for malformed ObjectIds (`errorHandler` maps Mongoose `CastError` to 404).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerErrorResponse": {
        "description": "Unhandled server error. `message` is the raw error text, so treat it as diagnostic only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "listUsers",
        "summary": "List users",
        "description": "Returns users (without passwords), newest first. `count` reflects the number of records returned in this response, so for the filtered non-admin cases `count === data.length`.\n\n**Who sees what** — the response is role-dependent, which is unusual for this API and worth stating plainly:\n\n| Caller role | Result |\n| --- | --- |\n| `admin` | Every user matching the optional `role` filter. |\n| `staff` | Forced to `role=staff` **and** `isActive=true`. Any other `role` value is ignored. |\n| `resident` | Same as `staff` — only active staff members, used to populate \"contact staff\" pickers. |\n\nA resident additionally gets **403** if they explicitly request `?role=` something other than `staff`.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleQuery"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Matching users. The set depends on the caller's role — see the description.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 4,
                      "description": "Always equals the length of `data` (this endpoint does not paginate)."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "post": {
        "operationId": "createUser",
        "summary": "Create a user",
        "description": "Creates an account. The caller supplies only the identity fields; the server generates the rest:\n\n- `username` — derived from the name (`firstname.lastname`), de-duplicated with a numeric suffix.\n- `password` — a cryptographically random temporary password.\n- `mustChangePassword` — set to `true`.\n\n**The temporary password is never included in the response.** Read the returned `message` and deliver the credential to the user through a secure channel. The password itself is only written to the server log, so if it is lost the admin must use `PUT /users/{id}/reset-password` to issue a new one.\n\n`specialization` is mandatory for `staff` accounts and forced to `null` otherwise. `exportSection` is forced to `all` for admins and `null` for residents.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "The user was created. **No password is returned** — `data` contains the identity fields only, and the `message` explains how to deliver the temporary password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "message": {
                      "type": "string",
                      "example": "User created. Deliver the temporary password through a secure channel; it must be changed on first login."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
    "/users/me/profile": {
      "put": {
        "operationId": "updateMyProfile",
        "summary": "Update my own profile",
        "description": "Lets any authenticated user edit their own profile. This is the only way to change your own password, and the only way to clear `mustChangePassword`.\n\n- Changing `name` also regenerates `username` (only when the name actually changed, so repeated saves do not inflate the username).\n- To change the password, send **both** `currentPassword` and `newPassword`. A missing or wrong `currentPassword` returns 400, as does a `newPassword` that is not 8-64 characters and does not include uppercase, lowercase, a number, and a special character.\n- Omitting `currentPassword`/`newPassword` leaves the password untouched.\n\nAdmins should prefer `PUT /users/{id}` for editing *other* people.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMyProfileRequest"
              }
            }
          },
          "description": "Omit every field to leave the profile unchanged."
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The profile was updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    },
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "getUserById",
        "summary": "Get a user by ID",
        "description": "Returns a single user without their password.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "put": {
        "operationId": "updateUser",
        "summary": "Update a user",
        "description": "Admin-only partial update of another account. Only the keys present in the body are applied.\n\n**Last-admin protection:** the server refuses (400) to deactivate or demote the final active admin, since that would leave the system with no way to manage users. Create a second admin first.\n\nPromoting someone to `staff` requires a `specialization`, and demoting away from `staff` clears it. `exportSection` is normalised against the resulting role: admins are forced to `all`, residents to `null`; for staff it is honoured as sent. Omitting `exportSection` still applies that normalisation, so demoting a staff member to resident also revokes their export rights.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "delete": {
        "operationId": "deactivateUser",
        "summary": "Deactivate (soft-delete) a user",
        "description": "**This is a soft delete.** The account is kept with `isActive=false` so historical references stay meaningful, and the session is invalidated immediately because `protect` rejects inactive users.\n\nTo avoid dangling references, the server also cascades:\n\n- deletes the user's `resident_houses` links;\n- clears `owner`/`tenant` on any houses they held;\n- unsets `submittedBy`, `assignedTo` and `resolvedBy` on their complaints;\n- unsets `paidBy`, `submittedBy` and `verifiedBy` on their dues;\n- deletes their notifications;\n- removes their votes from poll options and clears `createdBy` on their polls;\n- clears `createdBy` on their notices.\n\n**Guard rails:** you cannot deactivate your own account (400), and you cannot deactivate the last active admin (400). Use `isActive: true` via `PUT /users/{id}` to reactivate someone.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The user was deactivated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
...
  }
}